Add Dockerfile for development image
Rather than install node and npm locally I'm going to try and do all my
Elm stuff in Docker images.

To use the image, I'm running:

    $ docker image build -t elm-bingo .
    $ docker container run -it --rm elm-bingo

The pragstudio course is written to work with Elm 0.18 [1], though
things have moved on and the latest version of Elm is 0.19. I'm using
node 7.2 because the elm 0.18 package doesn't install cleanly with the
version of npm in the 10.x or 12.x node images, and because maport got
it going with 7.2 (see his Dockerfile [2]).

[1] https://online.pragmaticstudio.com/tour/courses/elm/steps/2
[2] https://github.com/maport/docker-elm/blob/master/elm/0.18/Dockerfile